# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is the Recipe API",
    "title": "Recipes API",
    "contact": {
      "name": "M Umer Masood",
      "url": "https://github.com/umermasood",
      "email": "umermasood.dev@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "listRecipes",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "description": "Create a new recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "newRecipe",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Tags",
            "name": "tags",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Ingredients",
            "name": "ingredients",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Instructions",
            "name": "instructions",
            "in": "query"
          },
          {
            "type": "string",
            "format": "date-time",
            "x-go-name": "PublishedAt",
            "name": "publishedAt",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/recipes/{id}": {
      "get": {
        "description": "Get one recipe",
        "produces": [
          "application/json"
        ],
        "operationId": "recipes",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "recipe ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "put": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "updateRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Invalid recipe ID"
          }
        }
      },
      "delete": {
        "description": "Delete an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Invalid recipe ID"
          }
        }
      }
    }
  }
}